% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{addCustom}
\alias{addCustom}
\title{addCustom function}
\usage{
addCustom(df, column, FUN = log, name = deparse(substitute(FUN)),
  end = TRUE, custom.name = NULL)
}
\arguments{
\item{data}{is a mts object}

\item{time_start}{is the starting period for trends}

\item{time_names}{is a vector of names for created trends}
}
\description{
This is a function allowing to create custom transformation for selected columns in a dataframe
}
\examples{
mtcars \%>\% addCustom(2:3, FUN = function(x) log(x), name = "lol")

transform_pmr(ts(mtcars), column = c("cyl", "disp"), method = "custom", end = TRUE, name = "r", FUN = function(x) {x * as.data.frame(ts(mtcars)[,1])},
custom.name = function(df, column, name) c(colnames(df),
                                          sapply(column, function(x) ifelse(length(grep("n", x)) > 0, gsub("n", "r", x), paste0(x, "_r")))))
}

