% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{addI}
\alias{addI}
\title{opis do zmiany}
\usage{
addI(df, column, FUN, name = deparse(substitute(FUN)), ...)
}
\arguments{
\item{df}{a data.frame object}

\item{column}{a character or numeric vector}

\item{FUN}{a transformation function to be used}
}
\description{
This function allows to use transformation function on several columns of mts or data.frame object
}

