% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{addT}
\alias{addT}
\title{general transformation expander function}
\usage{
addT(df, column, FUN, name = deparse(substitute(FUN)), ...)
}
\arguments{
\item{df}{a data.frame object}

\item{column}{a character or numeric vector}

\item{FUN}{a transformation function to be used}
}
\description{
This function allows to use transformation function on several columns of mts or data.frame object
}

