% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{addTrends}
\alias{addTrends}
\title{addTrends function}
\usage{
addTrends(data, time_start = 2000, time_names = c("time", "ln_time",
  "hyp_time", "sqrt_time"))
}
\arguments{
\item{data}{is a mts object}

\item{time_start}{is the starting period for trends}

\item{time_names}{is a vector of names for created trends}
}
\description{
This function adds 4 types of trends to the dataframe
}

