% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{best_models}
\alias{best_models}
\title{best_models function}
\usage{
best_models(object, specs = object$specs, signs = object$signs,
  type = object$type, method = "R2.DW")
}
\arguments{
\item{object}{an object from the getModels function}

\item{method}{a criteria used to determine the best model}

\item{a}{the number of best models to show based on R2.DW statistic}
}
\description{
Returns index of best models.
}

