% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{cton}
\alias{cton}
\title{transformation function}
\usage{
cton(x, base = 100)
}
\arguments{
\item{x}{a ts object}

\item{base}{}
}
\description{
n- nominal values, c- changes, i- chained index, r- relative index, p- percetange pointes or relative value
}

