% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{filterGap}
\alias{filterGap}
\title{filterGap function
adds hp filter gap calculated on logs}
\usage{
filterGap(df, column, lambda = 6.25)
}
\description{
filterGap function
adds hp filter gap calculated on logs
}

