% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{filterTrend}
\alias{filterTrend}
\title{filterTrend function
adds trend line from hp filter}
\usage{
filterTrend(df, column, lambda = 6.25)
}
\description{
filterTrend function
adds trend line from hp filter
}

