% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getModels}
\alias{getModels}
\title{getModels function}
\usage{
getModels(data, specs = setSpec(data), signs = setData(data))
}
\arguments{
\item{data}{test}

\item{specs}{test}

\item{signs}{test}
}
\description{
This is the main function to use in the ForIA package
The input is a data.frame, where the first column is the dependent variable. This function uses 2 optional functions as arguments
specs - which specifies all restrictions on the model selection
signs - which restricts the model space by restrictions on parameters
}
\examples{
getModels(cars)
}

