% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{ggplotRegression}
\alias{ggplotRegression}
\title{ggplotRegression}
\usage{
ggplotRegression(object, selection = 1:length(object$models))
}
\arguments{
\item{object}{an object from the getModels function}

\item{selection}{an index of the models to plot; works best with best_models function}

\item{method}{a criteria used to determine the best model}
}
\description{
returns wrapped plot of y vs fitted values
}

