% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{itov}
\alias{itov}
\title{transformation function}
\usage{
itov(x)
}
\arguments{
\item{x}{a ts object}
}
\description{
n- nominal values, c- changes, i- chained index, r- relative index, p- percetange pointes or relative value
}

