% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{lag_pmr}
\alias{lag_pmr}
\title{transformation expander for lag function}
\usage{
lag_pmr(df, column, lag = 1, ...)
}
\arguments{
\item{df}{a data.frame object}

\item{column}{a character or numeric vector}

\item{lag}{is the number of lags (may be negative)}
}
\description{
This function allows to use transformation function on several columns of mts or data.frame object
}

