% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{loadModels}
\alias{loadModels}
\title{loadModels function
loading list of models models from a text file}
\usage{
loadModels(...)
}
\description{
loadModels function
loading list of models models from a text file
}

