% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{merge_ts}
\alias{merge_ts}
\title{merge ts function}
\usage{
merge_ts(df_list)
}
\arguments{
\item{df_list}{is a list of mts objects}
}
\description{
This function merges mts objects from list of mts objects with thier column names
}

