% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{restrictModels2}
\alias{restrictModels2}
\title{restrictModels2 function}
\usage{
restrictModels2(object, specs = object$specs, signs = object$signs,
  type = object$type)
}
\arguments{
\item{object}{the object from getModels function}

\item{specs}{specs from the object, may be replaced by setSpecs output}

\item{signs}{signs from the object, may be replaced by setSigns output}
}
\description{
This function is used by the topModels functions when restrictions = TRUE. It sets the restrictions which vere specified in setData and setSpec functions
}

