% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{saveExcel}
\alias{saveExcel}
\title{saveExcel function
save excel accodring to Anrzej's format}
\usage{
saveExcel(object, elements, ...)
}
\description{
saveExcel function
save excel accodring to Anrzej's format
}

