% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveModels}
\alias{saveModels}
\title{saveModels function
saving list of models to a text file}
\usage{
saveModels(object, ...)
}
\arguments{
\item{object}{object from getModels function}
}
\description{
saveModels function
saving list of models to a text file
}

