% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{setData2}
\alias{setData2}
\title{setData3 function}
\usage{
setData2(data, cutCols = 1, positive = NA, negative = NA,
  must.have = NA, at.most = NA)
}
\arguments{
\item{data}{the data, used for colnames and calculating number of variables}

\item{positive}{represents vector of variables with positive signs}

\item{negative}{represents vector of variables with negative signs}

\item{must.have}{represents vector of variables that must be included at least one}

\item{at.most}{limits the amount of variables to m (option in setSpec function) in a model}
}
\description{
This function allows to set restrinctions on signs and number of variables included in the space of restricted models
This function is used by the getModels function and then transfered to some other funcions such as topModels
}

