% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setSpec}
\alias{setSpec}
\title{setSpec function
This function allows to set restrictions on some models characteristics when calling the restrictModels function}
\usage{
setSpec(data, T = dim(na.omit(data))[1], m = 1, nv.max = min(T - 2,
  length(colnames(data)[-1]), 3), alfa = 0.1, j.b = 0.05, b.p = 0.05,
  R.s = 0)
}
\arguments{
\item{data}{the data, used to set limit on the maximum amount of variables}

\item{T}{the length of the data, should not be changed by user if not necessary}

\item{m}{refers to the at.most parameter from setData function}

\item{nv.max}{the maximum number of variables in a model, (default = min(T-2, length(colnames(data)[-1]),3)) due to computation time}

\item{alfa}{significance level of the t-test}

\item{j.b}{p value for jb statistics}

\item{b.p}{p value for b.p statistics}

\item{R.s}{minimum accepted r squared adjusted}
}
\description{
setSpec function
This function allows to set restrictions on some models characteristics when calling the restrictModels function
}

