% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{topModels}
\alias{topModels}
\title{topModels function}
\usage{
topModels(object, a = 9, restrictions = TRUE, ...)
}
\arguments{
\item{object}{an object from the getModels function}

\item{a}{the number of best models to show based on R2.DW statistic}

\item{restrictions}{boolean specifying if restrictions from restrictModels should be applied}
}
\description{
Returns "a" best models. When restrictions is set TRUE, a restrictModels function is used to set restrictions on the model importance.
}

