% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{transform_pmr}
\alias{transform_pmr}
\title{transform.pmr function}
\usage{
transform_pmr(data, column, method = "ln", ...)
}
\arguments{
\item{data}{is an data.frame or mts object}

\item{column}{a character or numeric vector}

\item{method}{is the transformation method used}
}
\description{
This function wraps the transformation functions allowing for easy use with magrittrl package
}

