% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{addCustom}
\alias{addCustom}
\title{addCustom function}
\usage{
addCustom(df, column, FUN = log, name = deparse(substitute(FUN)),
  end = TRUE)
}
\arguments{
\item{data}{is a mts object}

\item{time_start}{is the starting period for trends}

\item{time_names}{is a vector of names for created trends}
}
\description{
This is a function allowing to create custom transformation for selected columns in a dataframe
}
\examples{
mtcars \%>\% addCustom(2:3, FUN = function(x) log(x), name = "lol")
}

