% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{getModels}
\alias{getModels}
\title{getModels function}
\usage{
getModels(data, type = "lm", specs = setSpec(data), signs = if (type ==
  "plm") setData(data[, -c(1, 2)]) else setData(data), ...)
}
\arguments{
\item{data}{test}

\item{specs}{test}

\item{signs}{test}
}
\description{
This is the main function to use in the ForIA package
The input is a data.frame, where the first column is the dependent variable. This function uses 2 optional functions as arguments
specs - which specifies all restrictions on the model selection
signs - which restricts the model space by restrictions on parameters
}
\examples{
getModels(cars)
}

