% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ForIA.R
\name{saveModels}
\alias{saveModels}
\title{saveModels function
saving list of models to a text file}
\usage{
saveModels(object, ...)
}
\arguments{
\item{object}{object from getModels function}
}
\description{
saveModels function
saving list of models to a text file
}

