% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setData}
\alias{setData}
\title{setData function}
\usage{
setData(data, positive = rep(NA, length(colnames(data)[-1])),
  negative = rep(NA, length(colnames(data)[-1])), must.have = rep(NA,
  length(colnames(data)[-1])), at.most = rep(NA, length(colnames(data)[-1])))
}
\description{
setData function
}

