% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setSpec}
\alias{setSpec}
\title{setSpec function}
\usage{
setSpec(data, T = dim(na.omit(data))[1], m = 1, nv.max = min(T - 2,
  length(colnames(data)[-1]), 3), alfa = 0.1, j.b = 0.05, b.p = 0.05,
  R.s = 0)
}
\description{
setSpec function
}

